scrapper/requirements.txt
beautifulsoup4==4.5.3
bs4==0.0.1
hurry.filesize==0.9
Naked==0.1.31
pycrypto==2.6.1
PyYAML==3.12
requests==2.12.4
selenium==3.0.2
shellescape==3.4.1
